feat: add files and configuration CI/CD

+++ src/main/java/br/ifrn/edu/jeferson/ecommerce/service/CategoriaService.java
@@ -3,6 +3,9 @@ package br.ifrn.edu.jeferson.ecommerce.service;
 import br.ifrn.edu.jeferson.ecommerce.domain.Categoria;
 import br.ifrn.edu.jeferson.ecommerce.domain.dtos.CategoriaRequestDTO;
 import br.ifrn.edu.jeferson.ecommerce.domain.dtos.CategoriaResponseDTO;
+import br.ifrn.edu.jeferson.ecommerce.exception.BusinessException;
+import br.ifrn.edu.jeferson.ecommerce.exception.ResourceNotFoundException;
+import br.ifrn.edu.jeferson.ecommerce.mapper.CategoriaMapper;
 import br.ifrn.edu.jeferson.ecommerce.repository.CategoriaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,51 +18,50 @@ public class CategoriaService {
     @Autowired
     private CategoriaRepository categoriaRepository;
 
-    public CategoriaResponseDTO salvar(CategoriaRequestDTO categoriaDto) {
+    @Autowired
+    private CategoriaMapper mapper;
+    @Autowired
+    private CategoriaMapper categoriaMapper;
 
-        Categoria categoria = new Categoria();
-        categoria.setNome(categoriaDto.getNome());
-        categoria.setDescricao(categoriaDto.getDescricao());
+    public CategoriaResponseDTO salvar(CategoriaRequestDTO categoriaDto) {
+        var categoria =  mapper.toEntity(categoriaDto);
+
+        if (categoriaRepository.existsByNome(categoria.getNome())) {
+            throw new BusinessException("Já existe uma categoria com esse nome");
+        }
+
         categoriaRepository.save(categoria);
-        return new CategoriaResponseDTO(categoria.getId(), categoria.getNome(), categoria.getDescricao());
+        return mapper.toResponseDTO(categoria);
     }
 
     public List<CategoriaResponseDTO> lista(){
         List<Categoria> categorias = categoriaRepository.findAll();
-        List<CategoriaResponseDTO> response = new ArrayList<>();
-
-        for (Categoria categoria : categorias) {
-            CategoriaResponseDTO categoriaDto = new CategoriaResponseDTO();
-            categoriaDto.setId(categoria.getId());
-            categoriaDto.setNome(categoria.getNome());
-            categoriaDto.setDescricao(categoria.getDescricao());
-            response.add(categoriaDto);
-        }
-
-        return response;
+        return mapper.toDTOList (categorias);
     }
 
     public void deletar(Long id) {
+        if (!categoriaRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Categoria não encontrada");
+        }
         categoriaRepository.deleteById(id);
     }
 
     public CategoriaResponseDTO atualizar(Long id, CategoriaRequestDTO categoriaDto) {
-        Categoria categoria = categoriaRepository.findById(id).orElse(null);
-        if(categoria != null) {
-            categoria.setNome(categoriaDto.getNome());
-            categoria.setDescricao(categoriaDto.getDescricao());
-            categoriaRepository.save(categoria);
-            return new CategoriaResponseDTO(categoria.getId(), categoria.getNome(), categoria.getDescricao());
+        Categoria categoria = categoriaRepository.findById(id).orElseThrow( () -> new ResourceNotFoundException("Categoria não encontrada"));
+
+        if (!categoria.getNome().equals(categoriaDto.getNome()) && categoriaRepository.existsByNome( categoriaDto.getNome()) ) {
+            throw  new BusinessException("Já existe uma categoria com esse nome");
         }
-        return null;
+
+        categoriaMapper.updateEntityFromDTO(categoriaDto, categoria);
+        var categoriaAlterada = categoriaRepository.save(categoria);
+
+        return categoriaMapper.toResponseDTO(categoriaAlterada);
     }
 
     public CategoriaResponseDTO buscarPorId(Long id) {
-        Categoria categoria = categoriaRepository.findById(id).orElse(null);
-        if(categoria != null) {
-            return new CategoriaResponseDTO(categoria.getId(), categoria.getNome(), categoria.getDescricao());
-        }
-        return null;
+        Categoria categoria = categoriaRepository.findById(id).orElseThrow( () -> new ResourceNotFoundException("Categoria não encontrada"));
+        return categoriaMapper.toResponseDTO(categoria);
     }
 
 }
